backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu==1.7.4",
    "fastapi>=0.118.0",
    "fuzzywuzzy>=0.18.0",
    "google-genai>=1.40.0",
    "gtts>=2.5.4",
    "ipython>=8.37.0",
    "numpy<2",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "twilio>=9.8.3",
    "uvicorn>=0.37.0",
]
